Increase the size of the banner image by 340px
This is to account for the decrease in size in the actual live demo.
The decrease in size happens because a fixed size of 310px locks the
size in HTML and CSS (because CSS is using max-width: 100%). The only
way to get around this issue is to increase the size in the HTML.